refactor: ACI-5128 one file per fix
